added file check and redo to easel run

+++ scripts/run_easel_getseqs.py
@@ -34,14 +34,22 @@ def run_easel(easelpath, databasepath, phmmerpath, keyfilepath, redo):
     filename = easeled_seq_formatter(keyfilepath)
     outpath = phmmerpath.joinpath(filename)
 
-    f = open(outpath, 'w+')
-    cmdargs = [f"{easelpath}/esl-sfetch",
-               "-f",
-               f"{databasepath}",
-               f"{keyfilepath}"] 
-
-    proc = subprocess.run(cmdargs, stdout=f)
-    f.close()
-
-    return proc
-
+    if not does_target_exist(keyfilepath, 'file'):
+        raise FileNotFoundError(f'KEYFILE MISSING: Could not find {keyfilepath}!')
+    elif does_target_exist(outpath, 'file') and redo == False:
+        print(f'Easel-fetched Fasta file: ({outpath}) already exists in {outpath.parent}')
+        return outpath
+    else:
+        try:
+            f = open(outpath, 'w+')
+            cmdargs = [f"{easelpath}/esl-sfetch",
+                       "-f",
+                       f"{databasepath}",
+                       f"{keyfilepath}"] 
+            proc = subprocess.run(cmdargs, stdout=f)
+            f.close()
+            if proc.returncode == 0:
+                print(f'Retrieved seqs in fasta file: {outpath}')
+            return outpath
+        except:
+            raise Exception('Could not extract seqs with easel')
